% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentation.R
\docType{data}
\name{ackflagg}
\alias{ackflagg}
\title{Code number and description}
\format{
A data frame with 3 variables:
\describe{
  \item{numeric code for fate of catch}
  \item{TEXT}{Description in Swedish}
}
}
\usage{
ackflagg
}
\description{
A dataset with commercially fished species. Each species is
described with a three character long code (MAF/FAO), english
common name and swedish common name.
}
\keyword{datasets}
