% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Read Excel-file and anonymize some columns}
\usage{
anonymize(
  x,
  columns = c("SIGNAL", "DISTRIKT", "BATNAMN", "PNR_DAT", "PNR_SLUT", "NAMN")
)
}
\arguments{
\item{x}{data.frame with catch data as requested from Hav}

\item{columns}{Columns to anonymize and remove from output}
}
\value{
Return a list with two tibbles named "anonymized_data" and "id2data".
In "anonymized_data" the columns specified in parameter "columns" are
replace with a single column "anon_id". The tibble "id2data" contains
the removed columns + the "anon_id" making it possible join the removed
columns back to the data if needed.
}
\description{
Function anonymize reads a file with (catch) data + data
about individual fishermen. This can make the file difficult to work
with in a GDPR safe way. The data can saved as two files a file with
the sensitive data columns replaced with one column with an
anonymous id and one file with the personal data and the
anon_id. If you save a file with personal data it should be stored in
a secure location with limited access.
}
\examples{
\dontrun{
result <- anonymize(catchdata)
}

}
