% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DAT2Date}
\alias{DAT2Date}
\title{HaV date to Date}
\usage{
DAT2Date(x)
}
\arguments{
\item{x}{integer a date in "HaV-format"}
}
\value{
Return the R Date converted from the input in "HaV-format"
}
\description{
The HaV date format is an integer in the format YYYYMMDD
This function returns the day as a R Date.
}
\examples{
DAT2Date(20200617)


}
\seealso{
[base::Dates] for Date class,

[DAT2year()] for year,
[DAT2month()] for month,
and [DAT2month()] for day.
}
