% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DAT2month}
\alias{DAT2month}
\title{HaV date to month}
\usage{
DAT2month(x)
}
\arguments{
\item{x}{integer a date in "HaV-format"}
}
\value{
Return the month part of a date in "HaV-format"
}
\description{
The HaV date format is an integer in the format YYYYMMDD
}
\details{
This function returns the month as an integer.
}
\examples{
DAT2month(20211224)

}
