% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DAT2year}
\alias{DAT2year}
\title{HaV date to year}
\usage{
DAT2year(x)
}
\arguments{
\item{x}{integer a date in "HaV-format"}
}
\value{
Return the year part of a date in "HaV-format"
}
\description{
The HaV date format is an integer in the format YYYYMMDD
}
\details{
This function returns the year as an integer.
}
\examples{
DAT2year(19610224)

}
