% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentation.R
\docType{data}
\name{fishing_area}
\alias{fishing_area}
\title{Table to map SUBDIV to FishingArea}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
fishing_area
}
\description{
\describe{
\item{SUBDIV}{character, ICES subdivisions in the Baltic sea}
\item{FishingArea}{character, corresponding fishing area used in Intercatch}
}
}
\keyword{datasets}
