% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentation.R
\docType{data}
\name{fleet_key}
\alias{fleet_key}
\title{Table to map HaV REDSKKOD to WGBAST metier}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 61 rows and 4 columns.
}
\usage{
fleet_key
}
\description{
\describe{
\item{metier}{character, metier used in Intercatch defined by WGBAST}
\item{gear}{character, gear name in Swedish defined by HaV}
\item{gear_code}{numeric, gear code defined by HaV}
\item{UnitEffort}{character, code for the unit for effort used for this gear type. Used in Intercatch (defined by WGBAST)}
}
}
\keyword{datasets}
