% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentation.R
\docType{data}
\name{gear}
\alias{gear}
\title{Names, code number and miscellaneous data of gear types}
\format{
A data frame with 10 variables:
\describe{
  \item{REDSKAP}{gear name in Swedish}
  \item{REDSKAPENG}{gear name in English}
  \item{REDSKKOD}{numeric code for gear}
  \item{AP}{Gear is "Passivt" or "Aktivt"}
  \item{EUKODNY}{Some other classification}
  \item{FR}{Gear is "Fast" or "Rörligt"}
  \item{FAOKODNY}{Some other classification}
}
}
\usage{
gear
}
\description{
A dataset describing fishing gear used in data files with catches reported
to HaV from commercial fishermen.
}
\keyword{datasets}
