% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_folder}
\alias{output_folder}
\title{Create (if needed) a folder to store files}
\usage{
output_folder(data_year, toplevel = "./import")
}
\arguments{
\item{data_year}{the year that data stored belongs to (catch year)}

\item{toplevel}{toplevel folder that will hold sub-folders for each year Default "./import"}
}
\value{
character the path of the folder

Returns a path to a folder that should hold data for a specific year. The year-subfolder is
placed under toplevel (default ./import). The folders are created if needed.
}
\description{
Utility function used to create a directory structure where you,
for example, store your results under a a folder for each year.
For example I save InterCatch import file in subfolders under "./import"
}
\examples{
odir <- output_folder(2019)

}
