% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slurp_xlsx}
\alias{slurp_xlsx}
\title{Read several Excel files and return the concatenated data}
\usage{
slurp_xlsx(fnames)
}
\arguments{
\item{fnames}{Character vector with one or more file namns}
}
\value{
a data frame
}
\description{
Read several Excel files and return the concatenated data
}
\examples{
\dontrun{
all_data <- slurp_xlsx(c("1.xlsx", "2.xlsx"))
}

}
