% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slurp_xlxs}
\alias{slurp_xlxs}
\title{Read several Excel files and return the concatenated data}
\usage{
slurp_xlxs(fnames)
}
\arguments{
\item{fnames}{Character vector with one or more file namns}
}
\value{

}
\description{
Read several Excel files and return the concatenated data
}
