% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-documentation.R
\docType{data}
\name{species}
\alias{species}
\title{Names, code number and miscellaneous data of gear types}
\format{
A data frame with 3 variables:
\describe{
  \item{MAFKOD}{a three charatcter long code (MAF/FAO), E.g. SAL and TRS}
  \item{Species}{English common name, e.g. Salmon and Trout}
  \item{Art}{Swedish common name, e.g. Lax and Öring}
}
}
\usage{
species
}
\description{
A dataset with commercially fished species. Each species is
described with a three character long code (MAF/FAO), english
common name and swedish common name.
}
\keyword{datasets}
